Add timer and cleanup listeners
Now activation listeners clean up after themselves. Delaypage is a
little big now and should probably be split apart. Uses
chrome.windows.getCurrent to evaluate whether tabs are still in
existence, and then the callback function closes or redirects the page
as needed. All of this is then used as a calllback for delayPageload,
and they rely on local variables defined within the delayPageLoad
context.

To do:

- Add options page

- Add unit tests

- Refactor delayPageLoad